+++ packages/terra-tabs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added
   * Added prop to support the focus traverse within the tab content.
+  * Fix for Active Key Style issue.
 
 ## 7.6.0 - (September 5, 2023)